move search_context
