+++ lib/appium_lib/android/android.rb
@@ -1,4 +1,3 @@
-require_relative 'search_context'
 require_relative 'command'
 require_relative 'device'
 
@@ -18,7 +17,6 @@ module Appium
     class Bridge
       def self.for(target)
         target.extend Appium::Android
-        target.extend Appium::Android::SearchContext
         target.extend Appium::Android::Device
       end
     end
