+++ lib/appium_lib/core/android.rb
@@ -1,0 +1,2 @@
+# loaded in common/driver.rb
+require_relative 'android/search_context'
