+++ lib/appium_lib/core/android_uiautomator2.rb
@@ -1,0 +1,1 @@
+# loaded in common/driver.rb
