+++ lib/appium_lib/core/common.rb
@@ -3,3 +3,6 @@ require_relative 'common/driver'
 require_relative 'common/error'
 require_relative 'common/log'
 require_relative 'common/wait'
+require_relative 'common/search_context'
+require_relative 'common/command'
+require_relative 'common/capabilities'
