+++ lib/appium_lib/core/common/capabilities.rb
