+++ lib/appium_lib/core/common/command.rb
