+++ lib/appium_lib/core/common/search_context.rb
@@ -1,0 +1,7 @@
+module Appium
+  module Core
+    module SearchContext
+      FINDERS = ::Selenium::WebDriver::SearchContext::FINDERS.merge(accessibility_id: 'accessibility id')
+    end
+  end
+end
