+++ lib/appium_lib/core/device.rb
@@ -354,7 +354,7 @@ module Appium
             end
 
             def _set_by_from_finders(how)
-              finders = ::Appium::Driver::SearchContext::FINDERS
+              finders = ::Appium::Core::SearchContext::FINDERS
               by = finders[how.to_sym]
               raise ArgumentError, "cannot find element by #{how.inspect}" unless by
               by
