+++ lib/appium_lib/core/ios.rb
@@ -1,0 +1,2 @@
+# loaded in common/driver.rb
+require_relative 'ios/search_context'
