+++ lib/appium_lib/core/ios_xcuitest.rb
@@ -1,0 +1,2 @@
+# loaded in common/driver.rb
+require_relative 'ios/xcuitest/search_context'
