--- lib/appium_lib/core/search_context.rb
@@ -1,7 +1,0 @@
-module Appium
-  module Core
-    module SearchContext
-      FINDERS = ::Selenium::WebDriver::SearchContext::FINDERS.merge(accessibility_id: 'accessibility id').freeze
-    end
-  end
-end
