+++ lib/appium_lib/ios/xcuitest/bridge.rb
@@ -6,10 +6,10 @@ module Appium
       class Bridge
         def self.for(target)
           target.extend Appium::Ios
-          target.extend Appium::Ios::SearchContext
+          # target.extend Appium::Ios::SearchContext
           target.extend Appium::Ios::Device
           target.extend Appium::Ios::Xcuitest
-          target.extend Appium::Ios::Xcuitest::SearchContext
+          # target.extend Appium::Ios::Xcuitest::SearchContext
           target.extend Appium::Ios::Xcuitest::Command
           target.extend Appium::Ios::Xcuitest::Helper
           target.extend Appium::Ios::Xcuitest::Gesture
